Added your name to the title element
Really awesome landing page for your website. The background image is a cool feature.
I added in your name to the <title> tag to make the page a little more identifiable on the tab list at the top a browser.
I would also consider creating a favicon so your website can have a unique icon show up on it's tab.